add ci, docker, and refactor the main.
