+++ .dockerignore
@@ -1,0 +1,13 @@
+.git
+.venv
+__pycache__
+*.pyc
+.mypy_cache
+.ruff_cache
+.pytest_cache
+tests/
+docs/
+docs-en/
+*.egg-info
+.anytask_session.json
+queue_db.json
