+++ .github/pull_request_template.md
@@ -1,0 +1,9 @@
+## Summary
+
+<!-- What does this PR do? -->
+
+## Test plan
+
+- [ ] `pytest tests/ -q` passes
+- [ ] `ruff check src/` passes
+- [ ] `mypy src/` passes
